Added extension specific profile functionality.
It's pretty much just a default profile for each extension. If no extension profile path is specified the default profile will be loaded. There's most definitely room for improvement in the code, but most scenarios should hopefully be accounted for. The extension specific profile functionality is only available for Wiimotes and is otherwise hidden.